Add avatar snapshot tests and refactor previews.
